Refactor landing page components and structure
