+++ src/app/page.tsx
@@ -1,33 +1,42 @@
 import dynamic from "next/dynamic";
-import RegisterSection from "@/components/RegisterSection/Register";
-import Footer from "@/components/Footer/Footer";
+import HeroSection from "@/components/landing-page/hero/Hero";
+import Footer from "@/components/landing-page/footer/Footer";
 
-const Welcome = dynamic(() => import("@/components/Welcome/Welcome"), {
-  loading: () => <div style={{ minHeight: 320 }} />,
-});
-const About = dynamic(() => import("@/components/About/About"), {
+const Welcome = dynamic(
+  () => import("@/components/landing-page/welcome/Welcome"),
+  {
+    loading: () => <div style={{ minHeight: 320 }} />,
+  },
+);
+const About = dynamic(() => import("@/components/landing-page/about/About"), {
   loading: () => <div style={{ minHeight: 320 }} />,
 });
 
 // Below-the-fold (deferred) sections
-const Gallery = dynamic(() => import("@/components/Gallery/Gallery"), {
-  loading: () => <div style={{ minHeight: 300 }} />,
-});
+const Gallery = dynamic(
+  () => import("@/components/landing-page/gallery/Gallery"),
+  {
+    loading: () => <div style={{ minHeight: 300 }} />,
+  },
+);
 const TeamCarousel = dynamic(
-  () => import("@/components/Carousel/TeamCarousel"),
+  () => import("@/components/landing-page/team/carousel/TeamCarousel"),
   {
     loading: () => <div style={{ minHeight: 260 }} />,
   },
 );
-const FAQ = dynamic(() => import("@/components/FAQ/FAQ"), {
+const FAQ = dynamic(() => import("@/components/landing-page/faq/FAQ"), {
   loading: () => <div style={{ minHeight: 260 }} />,
 });
-const Stats = dynamic(() => import("@/components/Stats/Stats"), {
+const Stats = dynamic(() => import("@/components/landing-page/stats/Stats"), {
   loading: () => <div style={{ minHeight: 260 }} />,
 });
-const Sponsor = dynamic(() => import("@/components/Sponsor/Sponsor"), {
-  loading: () => <div style={{ minHeight: 260 }} />,
-});
+const Sponsor = dynamic(
+  () => import("@/components/landing-page/sponsor/Sponsor"),
+  {
+    loading: () => <div style={{ minHeight: 260 }} />,
+  },
+);
 
 // Revalidate the home page (and thus the countdown) every hour
 export const revalidate = 3600; // seconds
@@ -36,8 +45,8 @@ export default function Home() {
   return (
     <>
       <main>
-        <section id="register">
-          <RegisterSection />
+        <section id="hero">
+          <HeroSection />
         </section>
 
         <section>
