+++ src/components/landing-page/about/About.tsx
