+++ src/components/landing-page/gallery/Gallery.tsx
