+++ src/components/landing-page/hero/background/DeferredGradientBackground.tsx
