+++ src/components/landing-page/hero/background/GradientBackground.tsx
