+++ src/components/landing-page/team/cards/TeamMemberCard.module.css
@@ -8,7 +8,7 @@
   height: 280px;
   border-radius: 0.5rem;
   overflow: visible;
-  background-image: url("../../assets/backgrounds/team-bg.webp");
+  background-image: url("../../../../assets/backgrounds/team-bg.webp");
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;
