+++ src/components/landing-page/team/carousel/carousel.module.css
